recommendation_engine/requirements.txt
pandas
numpy
scikit-learn
pymongo
redis
kafka-python
